Extract crm common methods to module
Extract crm common spec to separate file
