+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-04-23 15:05:37 +0300 using RuboCop version 0.67.2.
+# on 2019-04-23 15:35:07 +0300 using RuboCop version 0.67.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -15,16 +15,16 @@ Lint/UriEscapeUnescape:
 Metrics/AbcSize:
   Max: 23
 
-# Offense count: 17
+# Offense count: 18
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
-  Max: 1127
+  Max: 1078
 
 # Offense count: 4
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 541
+  Max: 515
 
 # Offense count: 1
 Metrics/CyclomaticComplexity:
@@ -46,22 +46,7 @@ Metrics/PerceivedComplexity:
 
 # Offense count: 69
 Naming/AccessorMethodName:
-  Exclude:
-    - 'lib/teamlab/modules/calendar.rb'
-    - 'lib/teamlab/modules/community/community_blogs.rb'
-    - 'lib/teamlab/modules/community/community_bookmarks.rb'
-    - 'lib/teamlab/modules/community/community_events.rb'
-    - 'lib/teamlab/modules/community/community_forums.rb'
-    - 'lib/teamlab/modules/community/community_wiki.rb'
-    - 'lib/teamlab/modules/crm.rb'
-    - 'lib/teamlab/modules/crm/crm_cases.rb'
-    - 'lib/teamlab/modules/files.rb'
-    - 'lib/teamlab/modules/group.rb'
-    - 'lib/teamlab/modules/mail.rb'
-    - 'lib/teamlab/modules/mail/mail_settings.rb'
-    - 'lib/teamlab/modules/people.rb'
-    - 'lib/teamlab/modules/project.rb'
-    - 'lib/teamlab/modules/settings.rb'
+  Enabled: false
 
 # Offense count: 1
 # Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist, MethodDefinitionMacros.
